// src/generated/wotlk/warlock.ts
export enum Warlock_Options_WeaponImbue {
    /**
     * @generated from protobuf enum value: NoWeaponImbue = 0;
     */
    NoWeaponImbue = 0,
    /**
     * @generated from protobuf enum value: GrandSpellstone = 1;
     */
    GrandSpellstone = 1,
    /**
     * @generated from protobuf enum value: GrandFirestone = 2;
     */
    GrandFirestone = 2
}